Add (pixel) image write instruction
Shader images, wee! This executes on the PBE unit, you pass it a PBE
descriptor pointed to by the texture state instead of a texture
descriptor. Same as with TODO.unkB1 (block image write) which also
executes on the PBE, but doesn't take value/coordinates/LOD? so has a
somewhat different encoding.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>