Decode image write block instruction
This instruction ("TODO.unkB1") is used to write out an entire block
from local memory into an image. Because it is block based and not pixel
based, in comparison to the regular image write instruction it works
even if the destination image is compressed. It is tailor fit for use in
the end-of-tile program, to blit tile memory to the framebuffer.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>
Co-developed-by: TellowKrinkle