Identify "doorbell" instructions
It's not clear exactly what these are for, but some sort of external
message-passing intrinsic. For now let's give the appropriate
generic name of "doorbell".

What's interesting is that, as used in the helper program, they are
always predicated by a special branch that is used to branch over the
intrinsic. Together the two instruction sequence does something like:

	if (should ring doorbell()) {
		doorbell(x);
	}

Details not understood :~)

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>